pyproject.toml
[project]
name = "cpanel-site-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.6.12"
dependencies = [
    "pillow>=8.4.0",
    "pyyaml>=6.0.1",
    "requests>=2.27.1",
    "scikit-image>=0.17.2",
    "selenium>=3.141.0",
]
